requirements.txt
tensorflow>=2.15
numpy>=1.24
matplotlib>=3.7
pillow>=10.0
scikit-learn>=1.3
